Return all results from the script.

+++ LogisticRegressionSample.m
@@ -1,6 +1,6 @@
 function [ params, time, gradient_value, error_bound, cc ] = LogisticRegressionSample( ftr, lbl, sample_rate, lambda, probability )
   num_examples = size(lbl, 1);
-  sample_size = num_examples * sample_rate;
+  sample_size = int32(num_examples * sample_rate);
   rand_perm = randperm(num_examples);
   fsample = ftr(rand_perm(1:sample_size), :);
   lsample = lbl(rand_perm(1:sample_size), :);
